% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather_request.R
\name{openweather_request}
\alias{openweather_request}
\title{Send a request to the OPEN WEATHER API}
\usage{
openweather_request(..., to_frame = TRUE, print_req = FALSE)
}
\arguments{
\item{...}{A series of named parameters to be used in the query.  Must be of the form
`param_key = "param_value"`.  Acceptable parameters are endpoint-specific.}

\item{to_frame}{A boolean value indicating whether or not the response
should be parsed and formatted as a data frame.  If `FALSE`,
a `response` object is returned and further processing can be done with
\code{\link[httr]{content}}.  Default is `TRUE`.}

\item{print_req}{A boolean value indicating whether or not the request
should be printed as well.  Useful for debugging.  Default is `FALSE`.}

\item{endpoint}{A string representing the OPEN WEATHER API endpoint of interest. _Required parameter._}
}
\value{
If `to_frame = TRUE`, a `tibble` containing the parsed response.
        If `to_frame = FALSE`, a `response` object returned directly from
        \code{\link[httr]{GET}}.
}
\description{
Send a general request to the OPEN WEATHER API by specifying an endpoint and a sequence
of parameters.  The `openweather_request()` function forms and submits a request to
a specified endpoint of the OPEN WEATHER API.  The return is either the `response`
object from \code{\link[httr]{GET}} or the response parsed as a `tibble`.
}
\section{API Documentation}{


[Open Weather Map API](https://openweathermap.org/api)
}

