% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather_series_observations.R
\name{openweather_series_observations}
\alias{openweather_series_observations}
\title{Get observations of a Open Weather Map Series}
\usage{
openweather_series_observations(id = NULL, aggregation_method = NULL,
  limit = NULL, sort_order = NULL, units = NULL, realtime_start = NULL,
  realtime_end = NULL)
}
\arguments{
\item{aggregation_method}{A string representing the aggregation method
used for frequency aggregation. This parameter has no affect is `frequency`
is not set. Possible values are:
* `"avg"` for average
* `"sum"` for sum
* `"eop"` for end of period value}

\item{limit}{An integer limit on the maximum number of results to return.
Defaults to `100000`, the maximum.}

\item{sort_order}{A string representing the order of the resulting series.
Possible values are: `"asc"` (default), and `"desc"`.}

\item{units}{A string indicating the data value transformation.
Defaults to `"F"`. Possible values are:

* `"C"` for Celcius}

\item{realtime_start}{A `Date` indicating the start of the real-time period.
Defaults to today's date.}

\item{realtime_end}{A `Date` indicating the end of the real-time period.
Defaults to today's date.}

\item{series_id}{A string ID fore the Open Weather series. _Required parameter_.}
}
\value{
A `tibble` object with observation dates and values.
}
\description{
Given the City ID, return observations of the series as a `tibble` object.
`openweather` is an alieas for `openweather_series_observations()`.
}
