% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather_series_observations.R
\name{openweather_series_observations}
\alias{openweather_series_observations}
\title{Get observations of a Open Weather Map Series}
\usage{
openweather_series_observations(id = NULL, type = NULL, cnt = NULL,
  units = NULL, aggregation_method = NULL, limit = NULL,
  sort_order = NULL, realtime_start = NULL, realtime_end = NULL)
}
\arguments{
\item{id}{A string ID for the Open Weather series. _Required parameter_.}

\item{type}{A string ID for the type of weather data to pull.  _Required parameter_.
* `"today"` for current days weather
* `"forecast"` for the 14 day forecast}

\item{units}{* `"i"` Imperial: Fahrenheit for temperature and miles/hour for wind speed
* `"k"` Default: Kelvin for temp and meter/sec for wind speed _Default_
* `"m"` Metric: Celsius for temp and meter/sec for wind speed}

\item{aggregation_method}{A string representing the aggregation method
used for frequency aggregation. This parameter has no affect is `frequency`
is not set. Possible values are:
* `"avg"` for average
* `"sum"` for sum
* `"eop"` for end of period value}

\item{limit}{An integer limit on the maximum number of results to return.
Defaults to `100000`, the maximum.}

\item{sort_order}{A string representing the order of the resulting series.
Possible values are: `"asc"` (default), and `"desc"`.}

\item{realtime_start}{A `Date` indicating the start of the real-time period.
Defaults to today's date.}

\item{realtime_end}{A `Date` indicating the end of the real-time period.
Defaults to today's date.}
}
\value{
A `tibble` object with observation dates and values.
}
\description{
Given the City ID, return observations of the series as a `tibble` object.
`openweather` is an alieas for `openweather_series_observations()`.
}
